fix bug introduced from buffer bits
